// src/models/finca.ts
export interface finca {
    apellidos: string;
    auxcodi: string;
    ciudad: string;
    contacto: string;
    direccion: string;
    email: string;
    entiDni: string;
    entiId: number;
    estado: string;
    fechregi: string;
    nombres: string;
    pais: string;
    phone: string;
    razosoci: string;
    tipo: string;
    tipoenti: string;
}